KS-16 rename: legacy -> item 도메인 구분으로 패키지 변경
