+++ src/main/java/com/kreamish/kream/item/repository/ItemRepository.java
@@ -1,6 +1,6 @@
-package com.kreamish.kream.legacy.repository;
+package com.kreamish.kream.item.repository;
 
-import com.kreamish.kream.legacy.entity.Item;
+import com.kreamish.kream.item.entity.Item;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface ItemRepository extends JpaRepository<Item, Long>, ItemQueryRepository {
